Refactor usedtokens to return actual token address
